Implement MVVM for Event Library Window

+++ DevNotePlay/ViewModels/EventTagViewModel.cs
@@ -1,20 +1,23 @@
 ﻿using LogApplication.Common.Config;
 using Player.Models;
 using Player.Services;
+using System;
 using System.Collections.ObjectModel;
+using System.ComponentModel;
 using System.Windows;
 
 namespace Player.ViewModels
 {
-    public class EventTagViewModel
+    public class EventTagViewModel: INotifyPropertyChanged
     {
         public ObservableCollection<EventTag> EventTags { get; set; }
 
         public RelayCommand UpdateCommand { get; set; }
         public RelayCommand DeleteCommand { get; set; }
+        public RelayCommand RefreshCommand { get; set; }
 
         private readonly string AppName;
-        private readonly EventTagService _eventTagService;
+        private EventTagService _eventTagService;
 
         public EventTagViewModel()
         {
@@ -23,12 +26,16 @@ namespace Player.ViewModels
 
             UpdateCommand = new RelayCommand(OnUpdate, CanUpdate);
             DeleteCommand = new RelayCommand(OnDelete, CanDelete);
+            RefreshCommand = new RelayCommand(OnRefresh);
 
             ConfigManager configManager = new ConfigManager();
             AppName = configManager.GetValue("AppName");
         }
 
         private EventTag _selectedEventTag;
+
+        public event PropertyChangedEventHandler PropertyChanged;
+
         public EventTag SelectedEvent
         {
             get
@@ -38,24 +45,22 @@ namespace Player.ViewModels
             set
             {
                 _selectedEventTag = value;
-                //UpdateCommand.RaiseCanExecuteChanged();
+                if (_selectedEventTag != null) _selectedEventTag.PropertyChanged += OnTargetUpdated;
+                UpdateCommand.RaiseCanExecuteChanged();
                 DeleteCommand.RaiseCanExecuteChanged();
             }
         }
 
         public void GetEventTags()
         {
-            //EventTagService _eventTagService = new EventTagService();
             EventTags = _eventTagService.GetEvents();
+            RaisePropertyChanged("EventTags");
         }
 
         private async void OnUpdate()
         {
-            //if (SelectedEvent == null)
-            //{
-            //    MessageBox.Show("Select an Event from the list first.", AppName, MessageBoxButton.OK, MessageBoxImage.Error);
-            //    return;
-            //}
+            var messageBoxResult = MessageBox.Show("Are you sure you want to update this item?", AppName, MessageBoxButton.YesNo);
+            if (messageBoxResult == MessageBoxResult.No) return;
 
             var eventTagService = new EventTagService();
             var result = await eventTagService.UpdateEventTag(SelectedEvent);
@@ -73,6 +78,9 @@ namespace Player.ViewModels
 
         private async void OnDelete()
         {
+            var messageBoxResult = MessageBox.Show("Are you sure you want to delete this item?", AppName, MessageBoxButton.YesNo);
+            if (messageBoxResult == MessageBoxResult.No) return;
+
             var eventTagService = new EventTagService();
             var result = await eventTagService.DeleteEventTag(SelectedEvent);
             //TODO: Make the error messages more meaningful by sending error details
@@ -87,15 +95,33 @@ namespace Player.ViewModels
             }
         }
 
+        private void OnRefresh()
+        {
+            _eventTagService = new EventTagService();
+            GetEventTags();
+        }
+
         private bool CanUpdate()
         {
-            //return SelectedEvent != null;
-            return true;
+            return SelectedEvent != null && SelectedEvent.IsValid();
         }
 
         private bool CanDelete()
         {
             return SelectedEvent != null;
+        }
+
+        private void OnTargetUpdated(Object sender, EventArgs e)
+        {
+            UpdateCommand.RaiseCanExecuteChanged();
+        }
+
+        private void RaisePropertyChanged(string property)
+        {
+            if (PropertyChanged != null)
+            {
+                PropertyChanged(this, new PropertyChangedEventArgs(property));
+            }
         }
     }
 }
